Move RunnableLanguage to core
